requirements.txt
Flask==2.2.5
gunicorn
tensorflow
numpy
Pillow
